Closes #90 renames IApp_* functions + minor improvements

+++ source/src/enet_encap/cpf.c
@@ -138,7 +138,7 @@ int NotifyConnectedCommonPacketFormat(EncapsulationData *received_data,
  */
 EipStatus CreateCommonPacketFormatStructure(
   const EipUint8 *data,
-  int data_length,
+  size_t data_length,
   CipCommonPacketFormatData *common_packet_format_data) {
 
   common_packet_format_data->address_info_item[0].type_id = 0;
